+++ features/support/env.rb
@@ -1,8 +1,5 @@
 require 'aruba/cucumber'
 require 'pry'
 
-# require 'geordi'
-# require 'cucumber/rspec/doubles'
-
 # Disables execution of Util.system! calls
 ENV['GEORDI_TESTING'] = 'true'