+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ### Compatible changes
 - Fixes [#70]: Make "console" and "server" commands ready for Rails 6
+- Added: CI testing with Travis
 
 ## 2.2.0 2019-02-28